Skip AsyncJavaTest on Nashorn that doesn't provide Promise
