+++ boot-script/src/test/java/net/java/html/boot/script/ScriptEngineCase.java
@@ -18,14 +18,17 @@
  */
 package net.java.html.boot.script;
 
+import java.io.IOException;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
+import net.java.html.js.JavaScriptBody;
 import org.netbeans.html.boot.spi.Fn;
 import org.netbeans.html.boot.impl.FnContext;
 import org.testng.IHookCallBack;
 import org.testng.IHookable;
 import org.testng.ITest;
 import org.testng.ITestResult;
+import org.testng.SkipException;
 import org.testng.annotations.Test;
 
 /**
@@ -50,6 +53,7 @@ public final class ScriptEngineCase implements ITest, IHookable {
 
     @Test
     public synchronized void executeTest() throws Throwable {
+        skipAsyncJavaTestWhenNoPromise();
         // BEGIN: net.java.html.boot.script.ScriptEngineCase#run
         Object instance = method.getDeclaringClass().newInstance();;
         for (int round = 0;; round++) {
@@ -72,4 +76,16 @@ public final class ScriptEngineCase implements ITest, IHookable {
         ihcb.runTestMethod(itr);
     }
 
+    private void skipAsyncJavaTestWhenNoPromise() throws IOException, SkipException {
+        if (method.getDeclaringClass().getSimpleName().equals("AsyncJavaTest")) {
+            try (var ctx = Fn.activate(p)) {
+                if (!typeOfPromise().equals("function")) {
+                    throw new SkipException("Skipping " + method.getName() + " no Promise found");
+                }
+            }
+        }
+    }
+
+    @JavaScriptBody(args = {}, body = "return typeof Promise")
+    private static native String typeOfPromise();
 }
